Generate the C++ vision client.
Note: this is currently bundled with the python simply because
it's easier to have the code gen work this way. We could consider
separating these, but I think we'd lose some ease in maintainance.

Also, there are some branches one must track to achieve this (as ever).

Steps to reproduce (if you've checked out this commit under checkoutroot):
1.  cd ~/mujin/jhbuildappcontroller/checkoutroot
2.  git checkout add-mujinvisioncontrollerclientcpp
3.  mujin_jhbuildcommon_initjhbuild.bash
4.  cd clientgenerators
5.  git checkout unit-test-merge-fix
6.  pip install oyaml mako
7.  cd ../visionmanager
8.  git checkout heman-fix-schema2
9.  jhbuild build -n mujinvisionmanagerpy
10. jhbuild buildone -n mujinclientgeneratorspy
11. cd ../clientgenerators
12. python devbin/mujin_clientgenerators_generateAllClients.py -vision -ow -langs h -langs cpp
13. jhbuild build -n mujinvisioncontrollerclientcpp

(Yes I had to re-indent everything once I typed out step 10.)
Here's some comments on some steps:

2)  This is needed for step 13. If you'd like not to do it, you still will need to
    checkout add-clientgenerators-vC for step 12.
6)  Note: this is for python2. Ideally, we'd be building mujinvisionmanagerpy
    in python3, but that's difficult (the whole dependency chain will need to
    build in py3).
11) This is just a convenient place to have a shell pointing to. But step 12
    assumes you did this.
12) You can omit the "-langs" arguments, but then you'll also clobber the python
    client which isn't ready (but doesn't error). Beware!
13) This is a sanity check.

I don't think there's much to say about the design. The specs need to be
tweaked to make sure the return values are correct and so that we respect
fireAndForget, but this is seems uncommon enough where I'll just use a
special case. It's interesting that so many of the parameters here were
mandatory. I've made some tweaks to be even smarter about "const T&" passing,
so will see if I need to update the sensorbridgeclient as well.